==> pandas-basic-functions/02_series_01.ipynb <==
# Create pandas series

import pandas as pd
import numpy as np

## 1. Using data structures

### 1.1. Lists

list1 = [11, 22, 33, 44, 55]

df1 = pd.Series(list1)

df1

### 1.2. Tuples

tuple1 = (11, 22, 33, 44, 55)

df2 = pd.Series(tuple1)

df2

### 1.3. Dictionaries

dict1 = {0: 11, 1: 22, 2: 33, 3: 44, 4: 55}

df3 = pd.Series(dict1)

df3

## 2. Using numpy

np1 = np.array([11, 22, 33, 44, 55])

df_np = pd.Series(np1)

df_np

## 3 Using scalars

scalar1 = pd.Series(11, index=[0, 1, 2, 3, 4])

scalar1
