==> pandas-basic-functions/03_series_02_direct.ipynb <==
# Create pandas series using direct method

import pandas as pd
import numpy as np

## 1. Lists

df_list = pd.Series([11, 22, 33, 44, 55])

df_list

## 2. Tuples

df_tuple = pd.Series((11, 22, 33, 44, 55))

df_tuple

## 3. Dictionaries

df_dict = pd.Series({0 : 11, 1 : 22, 2 : 33, 3 : 44, 4: 55})

df_dict

## 4. Numpy Arrays

df_np = pd.Series(np.array([11, 22, 33, 44, 55]))

df_np

## 5. Scalars

df_scalar = pd.Series(11, index=[0, 1, 2, 3, 4, 5])

df_scalar

## 6. Exercise

dict1 = {'A' : 100, 'B' : 200, 'C' : 300, 'D' : 400, 'E' : 500}

df_dict2 = pd.Series(dict1, index=dict1.keys())
# note that the 'index=dict1.keys()' is optional
# since pandas automatically assign the keys of the dictionary as the index of the series

df_dict2
