==> pandas-basic-functions/04_series_03_customized.ipynb <==
# Create pandas series - Customized indices & datatypes, access elements, and operations

import pandas as pd
import numpy as np

series1 = pd.Series([11, 22, 33, 44, 55])

series1

## 1. Change datatypes

series2 = pd.Series([11, 22, 33, 44, 55], index=['a', 'b', 'c', 'd', 'e'], dtype='float')

series2

series2_int = pd.Series([int(i) for i in series2])

series2_int

## 2. Access elements

series1[0]

series2[2:4]

## 3. Operations

### 3.1. Addition

s1 = pd.Series([1, 2, 3, 4, 5])
s2 = pd.Series([10, 20, 30, 40, 50])

s1 + s2

series1 + series2_int

### 3.2. Subtraction

s2 - s1

series2_int - series1

### 3.3. Multiplication

s1 * s2

series1 * series2_int

### 3.4. Division

s2 / s1

s1 / s2

series2_int / series1

### 3.5. Power

(s1 + s2) ** 2

(series1 + series2_int) ** 2 

## 4. Exercise

s3 = pd.Series([1, 2, 3, 4], index=[0, 1, 2, 3], dtype='int')
s4 = pd.Series([10, 20, 30, 40], index=[0, 1, 2, 3], dtype='int')

s3 // s4

s3 / s4
