==> pandas-basic-functions/05_dataframe_01.ipynb <==
# Intro to dataframes 

import pandas as pd
import numpy as np

## Create a dataframe

## 1. Using list

df_list = pd.DataFrame([11, 22, 33, 44, 55])

df_list

## 2. Using nested list

df_nested_list = pd.DataFrame([['Robert', 98], ['Alex', 97], ['Craig', 96]])

df_nested_list

nested_list = [['Robert', 98], ['Alex', 97], ['Craig', 96]]

df_nested_list_2 = pd.DataFrame(nested_list, columns=['Name', 'Score'])

df_nested_list_2

## 3. Using dictionary

df_dict = pd.DataFrame({'Name' : ['Robert', 'Alex', 'Craig'],
                        'Score': [98, 97, 96]})

df_dict

## 4. Using numpy array

array = np.array([[1, 2, 3],
                 [4, 5, 6],
                 [7, 8, 9]])

df_array = pd.DataFrame(array, columns=['a', 'b', 'c'], index=[0, 1, 2])

df_array

## 5. Using series

series_1 = pd.Series([11, 22, 33, 44, 55])

series_1

dataframe_1 = pd.DataFrame(series_1)

dataframe_1

series_2 = pd.Series(['v', 'w', 'x', 'y', 'z'])

dict_series = {'First' : series_1, 'Second' : series_2}

dataframe_2 = pd.DataFrame(dict_series)

dataframe_2

dict_students = {'Name' : ['Robert', 'Alex', 'Craig'],
                 'Score': [98, 97, 96]}

dataframe_3 = pd.DataFrame(dict_students, index=['Student 1', 'Student 2', 'Student 3'])

dataframe_3

## 6. Alter datatypes

dataframe_4 = pd.DataFrame(dict_students, columns=['Name', 'Score'], index=("Student 1", "Student 2", 'Student 3'))

dataframe_4

dataframe_4['Score Float'] = dataframe_4['Score'].astype('float')

dataframe_4

## 7. Exercise

city_temp_list = [['San Francisco', 30.1], ['Los Angeles', 32.2], ['Sacramento', '28.5']]

city_temp_df = pd.DataFrame(city_temp_list, columns=['City', 'Temperature'], index=[0, 1, 2])

city_temp_df
