==> pandas-basic-functions/06_dataframe_inspecting_data.ipynb <==
# Dataframes - Inspecting data

import pandas as pd
import numpy as np

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82]
}

df_student = pd.DataFrame(students_score, index=['s1', 's2', 's3', 's4', 's5', 's6', 's7'])

df_student

## 1. Dataset overview

df_student.shape

rows, columns = df_student.shape

print(f"Rows : {rows} \nColumns : {columns}")

df_student.head()

df_student.tail()

df_student.columns

df_student.info()

## 2. Accessing data

df_student['Name']

df_student[['Name', 'Score']]

df_student[:2]

df_student.iloc[0:2]

df_student.loc['s2', ['Name', 'Score']]

df_student.Name

df_student.Score

## 3. Basic calculations

df_student['Score'].max()

df_student[df_student['Score'] == df_student['Score'].max()]

df_student['Score'].min()

df_student[df_student['Score'] == df_student['Score'].max()]

df_student['Score'].mean()

df_student[df_student['Score'] >= df_student['Score'].mean()]

df_student.loc[df_student['Score'] >= df_student['Score'].mean()]

## 4. Basic statistics & quantities

df_student.describe()

df_student.value_counts()

df_student['Score'].value_counts()

df_student.count()

df_student['Score'].count()
