==> pandas-basic-functions/07_selection_bracket_dot.ipynb <==
# Column selection with bracket and dot notation

import pandas as pd
import numpy as np

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82]
}

df_students = pd.DataFrame(students_score, index=['s1', 's2', 's3', 's4', 's5', 's6', 's7'])

df_students

df_students['Name']

df_students.Name

df_students['Score']

df_students.Score

Always use bracket notation if there is any column name(s) with spaces, such as:

`df_students['Student Name']`

And if we want to select multiple columns, like:

`df_students[['Name', 'Score']]`

Otherwise we would get an error

students_score_2 = {
    'Student Name' : ['Pamela', 'Alessandro', 'Kryx', 'Vernon', 'Penny', 'Mifune', 'Autumn'],
    'Score': [98, 89, 99, 87, 90, 83, 82]
}

df_students_2 = pd.DataFrame(students_score_2)

df_students_2.student name

It returns an error.

Which won't be the case if we use bracket notation instead

df_students_2['Student Name']

Now, pay attention to this:

df_students.columns

df_students.head

Let's see another example. Let's say we want to create a dataframe with one or more of its column(s)
have the same names with python built in functions and attributes

shapes = {'shape': ['Round', 'Triangle', 'Rectangle', 'Square'],
          'head' : [1, 2, 3, 4]}

df_shapes = pd.DataFrame(shapes)

Now, if we access the column with bracket notation, we will get:

df_shapes.shape

See, instead of showing the data in the 'shape' column, the code above returns the shape
of the dataset because the 'df_shape' dataframe has a built-in function called .shape

Similar thing happens if we access the 'head' column with dot notation

df_shapes.head

It returns a method

Now, we compare if we use bracket ( [] ) notation instead

df_shapes['head']

df_shapes['shape']

Et voila! They actually return the values of each column
