==> pandas-basic-functions/08_loc_iloc.ipynb <==
# Data selection and subsetting with loc and iloc

## 1. loc

Access rows and columns by labels

Rules:

- loc:

    `dataframe.loc[Row, Column]`
    
    `dataframe.loc[Desired row(s), Desired column(s)]`

import pandas as pd
import numpy as np

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
    'Grades': ['AA', 'AB', 'AA', 'AC', 'AB', 'BA', 'BB']
}

df_students = pd.DataFrame(students_score)

df_students

The code below means 'select row with index 0 or select first row of all columns'

df_students.loc[0, :]

This code below means 'select row 0, 1, 2 of all columns'

df_students.loc[[0, 1, 2], :]

The code above can be simplified with:

df_students.loc[0:2, :]

Using column names

df_students.loc[:, 'Name']

df_students.loc[:, ['Name', 'Score']]

df_students.loc[:, 'Name':]

## 2. iloc

Access rows and columns by indices

Rules:

- iloc:

    Access a group of rows and columns by integer position(s)
    
    `dataframe.iloc[First Inclusive, Last Exclusive]`

df_students.iloc[0, :]

df_students.iloc[:, 0]

df_students.iloc[:, 0:2]

df_students.iloc[:, 0:]

df_students.iloc[[0, 1], [0, 1]]

df_students.iloc[0:3, :]
