==> pandas-basic-functions/09_loc_iloc_iat_at_ix.ipynb <==
# Selection and subsetting : loc, iloc, iat, at, ix

import pandas as pd

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
    'Grades': ['AA', 'AB', 'AA', 'AC', 'AB', 'BA', 'BB']
}

df_students = pd.DataFrame(students_score, index=['s1', 's2', 's3', 's4', 's5', 's6', 's7'])

df_students

## 1. loc

### 1.1. Name, Score, Grades of s1 and s2

df_students.loc[['s1', 's2'], :]

### 1.2. Name and Grades of s1, s3, s5

df_students.loc[['s1', 's3', 's5'], ['Name', 'Grades']]

### 1.3. Name, Score, and Grades of s1 to s7

df_students.loc[:, :]

df_students.loc['s1': 's7', :]

## 2. iloc

### 2.1. Name, Score, and Grades of s1 and s2

df_students

df_students.iloc[0:2, :]

df_students.iloc[[0, 1], :]

### 2.2. Name and Grades of s1, s3, s5

df_students.iloc[[0, 2, 4], [0, 2]]

### 2.3. Name, Score, and Grades of s1 to s7

df_students.iloc[:, :]

df_students.iloc[0:7, :]

df_students.iloc[:7, :]

## 3. iat

`iat` is a method that enables access a single value for a row or column pair by integer position

similar to `iloc`, both uses integer as arguments

use `iat` if you only need to get or set a single value in a dataframe or series

### 3.1. Grade of s2

df_students

df_students.loc['s2', 'Grades']

df_students.iat[1, 2]

### 3.2. Name of s5

df_students.loc['s5', 'Name']

df_students.iat[4, 0]

### 3.3. Set score of s5 to 99

df_students.loc['s5', 'Score'] = 99

df_students

df_students.iat[4, 1] = 99

df_students.iat[4, 2] = 'AA'

df_students

## 4. at

`at` accesses a single value for a row/column label pair

similar to loc, they both takes labels as arguments

Use `at` if you only need to get or set a single value in a dataframe or series

### 4.1. Grade of s2

df_students.loc['s2', 'Grades']

df_students.at['s2', 'Grades']

### 4.2. Name of s5

df_students.at['s5', 'Name']

### 4.3. Set Score of s5 to 99

df_students.at['s5', 'Score'] = 99

df_students.at['s5', 'Grades'] = 'AA'

df_students

df_students.loc['s5', ['Name', 'Score', 'Grades']]
