==> pandas-basic-functions/10_loc_iloc_boolean.ipynb <==
# loc, iloc with booleans

import pandas as pd

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
    'Grades': ['AA', 'AB', 'AA', 'AC', 'AB', 'BA', 'BB']
}

df_students = pd.DataFrame(students_score, index=['s1', 's2', 's3', 's4', 's5', 's6', 's7'])

df_students

## loc

### Access rows and columns by labels or boolean arrays

The size of the boolean arrays must be the same as the dataframe size

This is useful for filtering and subsetting data

`dataframe.loc[rows(string), columns(string)]`

df_students.loc[[False, False, False, False, False, False, False]]

df_students.loc[[True, False, False, True, False, False, True]]

with columns

df_students.loc[[False, True, False, False, False, True, False], [True, True, False]]

### 1. Name of s1 and s2

df_students.loc[[True, True, False, False, False, False, False], [True, False, False]]

### 2. Grades of s1 and s5

df_students.loc[[True, False, False, False, True, False, False], [False, False, True]]

### 3. Score of s4

df_students.loc[[False, False, False, True, False, False, False], [False, True, False]]

## iloc

We can apply this technique using iloc

df_students.iloc[[False, True, False, False, False, True, False], [True, True, True]]
