==> pandas-basic-functions/11_filtering_subsetting.ipynb <==
# Filtering and subsetting using conditionals without loc or iloc

import pandas as pd

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
    'Grades': ['AA', 'AB', 'AA', 'AC', 'AB', 'BA', 'BB']
}

df_students = pd.DataFrame(students_score, index=['s1', 's2', 's3', 's4', 's5', 's6', 's7'])

df_students

## 1. Without loc/iloc

### 1.1. Students with scores over 90

df_students[df_students['Score'] > 90]

### 1.2. Student(s) who has/have AB

df_students[df_students['Grades'] == 'AB']

### 1.3. Student(s) with scores over 90 and AB

df_students[(df_students['Score'] > 90) & (df_students['Grades'] == 'AB')]

Using `or` operator

df_students[(df_students['Grades'] == 'AB') | (df_students['Score'] > 90)]

### 1.4. Name of students with score over 90

df_students[df_students['Score'] > 90]['Name']

### 1.5. Name of students with AB

df_students[df_students['Grades'] == 'AB']['Name']

## 2. With loc

### 2.1. Students with score over 95 using `loc[]`

df_students.loc[df_students['Score'] > 95]

### 2.2. Students with score over 95 and AB using `loc[]`

df_students.loc[(df_students['Score'] > 95) & (df_students['Grades'] == 'AB')]

## 3. with iloc

### 3.1. Students with score over 90 using `iloc[]`

df_students.iloc[:, 1] > 90

df_students[df_students.iloc[:, 1] > 90]

### 3.2. Students with score over 90 and AB using `iloc[]`

df_students.iloc[:, 1] > 80

df_students.iloc[:, 2] == 'AB'

Combined, we get:

df_students[(df_students.iloc[:, 1] > 80) & (df_students.iloc[:, 2] == 'AB')]

### 3.3. Name of students with AB

df_students[df_students.iloc[:, 2] == 'AB']['Name']
