==> pandas-basic-functions/12_add_drop_columns.ipynb <==
# Add/drop columns to/from dataframe

import pandas as pd

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
}

df_students = pd.DataFrame(students_score)

df_students

## 1. List

#  'Grade': ['AA', 'AB', 'AA', 'AC', 'AB', 'BA', 'BB'] 

grade = ['AA', 'AB', 'AA', 'AC', 'AB', 'BA', 'BB']

df_students['Grade'] = grade

df_students

### 1.1. Drop the column(s)

df_students.drop(['Score', 'Grade'], axis=1, inplace=True)

df_students

## 2. Pandas series

grade_series = pd.Series(['AA', 'AB', 'AA', 'AC', 'AB', 'BA', 'BB'])

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
}

df_students = pd.DataFrame(students_score)

df_students

df_students['Grade'] = grade_series

df_students

## 3. Dictionary

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
}

df_students = pd.DataFrame(students_score)

df_students

dictio = {
    'Austin' : 'Paul',
    'Albany' : 'Aaron',
    'Nashville' : 'Krista',
    'Boston' : 'Varonica',
    'Knoxville' : 'Paxton',
    'Kansas City' : 'Madison',
    'Toledo' : 'Aurora' 
}

df_students['City'] = dictio.keys()

df_students

## 4. Insert a column(s)

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
}

df_students = pd.DataFrame(students_score)

df_students

grade_series = pd.Series(['AA', 'AB', 'AA', 'AC', 'AB', 'BA', 'BB'])

# Insert grade_series values at position 2
df_students.insert(2, 'Grade', grade_series)

df_students

## 5. Add a column from existing columns 

df_students['Name-Grade'] = df_students['Name'] + ', ' + df_students['Grade']

df_students

## 6. Assign

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
}

df_students = pd.DataFrame(students_score)

df_students

grade_series = pd.Series(['AA', 'AB', 'AA', 'AC', 'AB', 'BA', 'BB'])

state_series = pd.Series(['TX', 'NY', 'TN', 'MA', 'TN', 'MO', 'OH'])

df_students_2 = df_students.assign(Grade=grade_series, State=state_series)

df_students_2

df_students.assign(Score1=lambda x: x.Score + 1)

df_students.assign(Scoremin5=df_students['Score'] - 5)

df_students

temp_cities = pd.DataFrame({'temp_c' : [17.0, 25.0]},
                           index=['Portland', 'Berkeley'])

temp_cities

temp_cities.assign(temp_f = lambda x: x.temp_c * 9 / 5 + 32)

temp_cities.assign(temp_f = temp_cities['temp_c'] * 9 / 5 + 32)
