==> pandas-basic-functions/13_add_remove_rows.ipynb <==
# # Add/remove column(s) to/from dataframe

import pandas as pd

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
}

df_students = pd.DataFrame(students_score)

df_students

## 1. Using concat

df_students_2 = pd.DataFrame({"Name" : "Annika", "Score" : [99]})

df_students_3 = pd.concat([df_students, df_students_2], ignore_index=True)

df_students_3

## 2. Using Append

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
}

df_students_1 = pd.DataFrame(students_score)

df_students_1


row_student_add = {"Name" : "Robert", "Score" : [90]}

df_students_1 = df_students.append(row_student_add, ignore_index=True)

## 3. Using iloc

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
}

df_students_1 = pd.DataFrame(students_score)

df_students_1

df_students_1.iloc[0] = ['Archie', 89]

df_students_1

## 4. Using loc

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
}

df_students_1 = pd.DataFrame(students_score)

df_students_1

df_students_1.loc[2] = ['Robert', 90]

df_students_1

## 5. Remove rows

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
}

df_students_1 = pd.DataFrame(students_score)

df_students_1

df_students_1.drop([0, 1], inplace=True)
df_students_1.reset_index(drop=True)

Useful method

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
}

df_students_1 = pd.DataFrame(students_score)

df_students_1

df_students_1_under_90 = df_students_1[df_students_1['Score'] < 90]

df_students_1_under_90.index

df_students_1.drop(df_students_1_under_90.index)

One liner equavalent

df_students_1.drop(df_students_1[df_students_1['Score'] < 90].index)
