==> pandas-basic-functions/14_to_csv.ipynb <==
# Create a CSV file from pandas dataframe

import pandas as pd

students_score = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82],
    'Grades':['AA','AB','AA','AB','AC','BA','BB']
}

df_student = pd.DataFrame(students_score)

df_student

df_student.to_csv('students_score.csv')

pwd

df_student.to_csv('students_score_no_index.csv', index=False)

df_student_2 = pd.read_csv("students_score_no_index.csv")

df_student_2
