==> pandas-basic-functions/15_read.ipynb <==
# Read CSV, excel and TSV files from any directory

import pandas as pd

## 1. Comma separated values (CSV)

df_students = pd.read_csv("./datasets/students_score_no_index.csv")

df_students

## 2. Excel files

df_iris = pd.read_excel("./datasets/iris.xlsx", sheet_name="Sheet1")

df_iris.head()

df_iris.loc[:2, ['SepalLengthCm']]

## 3. Tab separated values (TSV)

review_comments = pd.read_table("./datasets/Restaurant_Reviews.tsv")

review_comments.head()

## 4. Read TSVs with `read_csv()`

reviews_csv = pd.read_csv("./datasets/Restaurant_Reviews.tsv", sep="\t")

reviews_csv.head()
