==> pandas-basic-functions/16_read_csv_nrows.ipynb <==
# `read_csv()` - `nrows` parameter

`nrows` parameter is used when we want to read samples of large datasets.
Basically it enables us to read number of rows of data from the dataset

import pandas as pd

df_student_sample = pd.read_csv("./datasets/students_score_no_index.csv", nrows=3)

df_student_sample
