==> pandas-basic-functions/17_read_csv_skiprows.ipynb <==
# `read_csv()` - `skiprows` parameter

import pandas as pd

Normal read

df_student_score = pd.read_csv("./datasets/students_score_no_index.csv")

df_student_score

## 1. Skipping rows from first (0th) position

Specifiy how many row(s) that we want to skip from the dataset

df_student_score = pd.read_csv("./datasets/students_score_no_index.csv", skiprows=2)
# means we skip first 2 rows

df_student_score

See the difference?

<b> * Always keep in mind that when we only use one integer as the value of skiprow parameter, it means we also skip the header of the dataset</b>

## 2. Skipping rows from a specific/custom position

<b> * Note that when we use list as the skiprow parameter, we won't skip the header of the dataset</b>

df_student_score = pd.read_csv("./datasets/students_score_no_index.csv", skiprows=[2, 4, 6])
# means we skip second, fourth, and fifth rows

df_student_score

## 3. Skipping rows using list comprehension

df_student_score = pd.read_csv("./datasets/students_score_no_index.csv", skiprows=[i for i in range(1, 3)])
# means we skip first row and second row
# because range(1, 3) means take 1 but do not take 3
# in other words range(1, 3) equals 1 and 2
# because (1, 3) means first inclusive (1) and last exclusive (3)

df_student_score

# skip first, third, and fifth row looks like:
df_student_score = pd.read_csv("./datasets/students_score_no_index.csv", skiprows=[i for i in [2, 4, 6]])
# means we skip second, fourth, and fifth rows

df_student_score

## 4. Skipping rows using conditions

Say we want to use a function that return a boolean value if a number is divisible by 3

def div_three(x):
    if x % 3 == 0:
        return True
    else:
        False

Now we pass the function using lambda as a value for the skiprow parameter, so that we may skip whichever row position that is divisible by three. In other words, we skip every third row

df_student_score = pd.read_csv("./datasets/students_score_no_index.csv", skiprows=lambda x: div_three(x))

df_student_score

## 5. Using `skipfooter` parameter

Skip number of rows from the end row of the dataset

df_student_score = pd.read_csv("./datasets/students_score_no_index.csv", skipfooter=3)

df_student_score

Skipped last three rows

# Use engine parameter to ignore the warning
df_student_score = pd.read_csv("./datasets/students_score_no_index.csv", skipfooter=3, engine='python')

df_student_score
