==> pandas-basic-functions/18_read_csv_usecols.ipynb <==
# `read_csv()` - `usecols` parameter

Load a dataset with specific column(s)

import pandas as pd

## 1. Using column names

df_students = pd.read_csv("./datasets/students_score_no_index.csv", usecols=["Name", "Score"])

df_students

## 2. Using column index

df_students = pd.read_csv("./datasets/students_score_no_index.csv", usecols=[0, 1])

df_students

## 3. Using column names as a list

# Create a list of column names
column_names = ['Name', 'Score']

df_students = pd.read_csv("./datasets/students_score_no_index.csv", usecols=column_names)

df_students

## 4. Use col_labels

df_students_full = pd.read_csv("./datasets/students.csv")

df_students_full

df_students_full_2 = pd.read_csv("./datasets/students.csv", usecols=lambda x: x.upper() in ['City', 'State', 'State Name'])

df_students_full_2
