==> pandas-basic-functions/19_read_csv_indexing.ipynb <==
# `read_csv()` indexing

import pandas as pd

## 1. `index_col`

df_score = pd.read_csv("./datasets/students_score_no_index.csv", index_col="Grades")
# or we can pass integers to index_col
# df_score = pd.read_csv("./datasets/students_score_no_index.csv", index_col=0)

df_score

## 2. `set_index()`

df_score = pd.read_csv("./datasets/students_score_no_index.csv")

df_score.set_index('Score')

## 3. Multi-indexing

df_score = pd.read_csv("./datasets/students_score_no_index.csv", index_col=["Name", "Score"])

df_score

## 4. Reset index

df_score.reset_index()

# to reset the index and save it to the dataframe 
df_score = df_score.reset_index()

df_score
