==> pandas-basic-functions/20_read_csv_header.ipynb <==
# `read_csv()` header parameter

import pandas as pd

## 1. Normal read

df_student = pd.read_csv("./datasets/students_score_no_header.csv")

df_student

## 2. Using `header=1`

df_student = pd.read_csv("./datasets/students_score_no_header.csv", header=1)

df_student

## 3. Using `header=None`

df_student = pd.read_csv("./datasets/students_score_no_header.csv", header=None)

df_student
