==> pandas-basic-functions/21_read_csv_names.ipynb <==
# `read_csv()` names parameter

import pandas as pd

## 1. Basic read

df_student_nh = pd.read_csv("./datasets/students_score_nh.csv")

df_student_nh

## 2. Using list of names directly

df_student_h = pd.read_csv("./datasets/students_score_nh.csv", names=["Name", "Score", "Grades"])

df_student_h

## 3. Using list of names from a variable

cols = ["Name", "Score", "Grades"]

df_student_col = pd.read_csv("./datasets/students_score_nh.csv", names = cols)

df_student_col

## 4. Combine with header parameter

df_student_head = pd.read_csv("./datasets/students_score_nh.csv", names=['Col A', 'Col B', 'Col C'], header=0)

df_student_head


## 5. Duplicate names are not allowed

# If we pass duplicates to that particular names parameter, it will genereate an error
# So instead of this
# df_student_dup = pd.read_csv("./datasets/students_score_nh.csv", names=['Col A', 'Col A', 'Col C'], header=0)

# We do this
df_student_dup = pd.read_csv("./datasets/students_score_nh.csv", names=['Col A', 'Col B', 'Col C'], header=0)


df_student_dup
