==> pandas-basic-functions/23_read_csv_dtype.ipynb <==
# `read_csv` dtype parameter

Change dtype of a column while reading a dataset

import pandas as pd

## 1. Basic read

df_student = pd.read_csv("./datasets/students_score_no_index.csv")

df_student

## 2. Using dtype parameter

df_student_f32 = pd.read_csv("./datasets/students_score_no_index.csv", dtype={'Score' : 'float32'})

df_student_f32

## 3. Multiple conversion

df_student_mult = pd.read_csv("./datasets/students_score_exp.csv", dtype={'Score' : 'float32', 'Queue' : 'float64', 'Order' : 'int32'})

df_student_mult
