==> pandas-basic-functions/24_read_csv_true_values_false_values.ipynb <==
# `read_csv` `true_values` & `false_values` parameter

Converting custom true and false values to boolean (True / False)

import pandas as pd

## 1. Basic read

df_std_att = pd.read_csv("./datasets/students_score_att.csv")

df_std_att

## 2. Read with `true_values` and `false_values`

df_std_att_bool = pd.read_csv("./datasets/students_score_att.csv", true_values=["Yes"], false_values=["No", "Not"])

df_std_att_bool

## 3. Find how many students were absent

# length of the data subtracted by sum of true values in attended column
len(df_std_att_bool) - df_std_att_bool["Attended"].sum()
