==> pandas-basic-functions/26_read_csv_large.ipynb <==
# Reading large CSV files

import pandas as pd

## 1. Basic read

df_iris = pd.read_csv("./datasets/iris.csv")

df_iris

## 2. Using chunksize

df_iris_chunk = pd.read_csv("./datasets/iris.csv", chunksize=10)

df_iris_chunk

for chunk in df_iris_chunk:
    print(chunk)

df_iris_chunk = pd.read_csv("./datasets/iris.csv", chunksize=10)

chunk_len = 0

for chunk in df_iris_chunk:
    chunk_len += len(chunk)
    
print(chunk_len)

## 3. Combine the chunks in to one dataframe using `pd.concat()`

df_iris_chunk = pd.read_csv("./datasets/iris.csv", chunksize=10)
df_iris_comb = pd.concat(df_iris_chunk, ignore_index=True)

df_iris_comb
