==> pandas-basic-functions/27_inplace.ipynb <==
# Inplace parameter

import pandas as pd

df_student_att = pd.read_csv("./datasets/students_score_att.csv")

df_student_att

## 1. Dropping a column without inplace

df_student_att.drop("Grades", axis=1)

Now we show the data again

df_student_att

Notice that the data hasn't changed, even we just dropped a column

To actually change the data we add inplace parameter set to True

So that the dataframe variable `df_student_att` actually changed based on the operations we perform

## 2. Dropping a column using inplace parameter

df_student_att.drop("Grades", axis=1, inplace=True)

df_student_att

Now that we dropped the "Grades" column and set the inplace parameter to True, the dataframe `df_student_att` has changed permanently

# The method
# df_student_att.drop("Grades", axis=1, inplace=True)
# is similar to
# df_student_att = df_student_att.drop("Grades", axis=1)
