==> pandas-basic-functions/28_na_isnull_notnull.ipynb <==
# Handling missing values - Isnull & notnull

import pandas as pd

df_na = pd.read_csv("./datasets/df_na.csv")

df_na

## 1. Find total number of missing values using `isnull()`

df_na.isnull()

df_na.isnull().sum()

df_na.isnull().sum().sum()

## 2. Using `isna()`

df_na.isna()

df_na.isna().sum().sum()

## 3. `notnull()`

Not null is exactly the opposite of isnull which is a method to find non null values

df_na.notnull()

df_na.notnull().sum()

## 4. Checking null/non null values in a column

type(df_na["Score"])

df_na[df_na["Score"].isnull()]

df_na[df_na["Score"].notnull()]
