==> pandas-basic-functions/29_na_values.ipynb <==
# Handling missing values - na_values

import pandas as pd

df_na_nav = pd.read_csv("./datasets/df_na_nav.csv")

df_na_nav

## Handling the wrong missing values from read

df_na_read = pd.read_csv("./datasets/df_na_nav.csv", na_values='not available')

df_na_read

As you can see 'not available' is converted to NaN

If you want to replace multiple unknown missing values with NaN, just pass the values as a list like below

# df_na_read = pd.read_csv("./datasets/df_na_nav.csv", na_values=['not available', 'wrong data', 'na values'])

## Handling missing values only in a particular column

pass the values as a dictionary with the key is the name of the column and the values are the name of the wrong values

df_na_dict = pd.read_csv("./datasets/df_na_nav.csv", na_values={'Grades' : 'not available'})

df_na_dict
