==> pandas-basic-functions/30_na_filter_keep_default_na.ipynb <==
# `na_filter` & `keep_default_na`

import pandas as pd

df_student = pd.read_csv("./datasets/df_na_nav.csv", na_values="not available")

df_student

df_student = pd.read_csv("./datasets/df_na_nav.csv", na_values="not available", keep_default_na=False)

df_student

df_student = pd.read_csv("./datasets/df_na_nav.csv", na_values="not available", na_filter=True)

df_student
