==> pandas-basic-functions/31_drop_na.ipynb <==
# Dropping missing values with `drop_na`

`data.dropna(axis=0, how='any', tresh=None, subset=None, inplace=False)`

1. axis (0 or 1)

    use 0 or 'index' to drop rows that contain missing values, or <br>
    use 1 or 'column names' to drop columns that contain missing values <br>

2. Drop rows or columns that have at least one missing values <br>

    'any' : if any NA values are present <br>
    'all' : if all values are NA <br>

3. Set some treshold value to drop missing values <br>

    tresh : int, optional

4. Drop rows of selected columns

    subset : define columns that have missing values

5. Drop rows or columns temporarily or permanently

    inplace : boole (True of False)





import pandas as pd

df_student = pd.read_csv("./datasets/df_m.csv")

df_student

## any

df_student.dropna(axis=0, how='any')

## Drop columns where at least one value is missing

df_student.dropna(axis=1, how='any')

## Drop rows where all values are missing

df_student.dropna(axis=0, how='all')

## Drop the columns where all values are missing

df_student.dropna(axis='columns', how='all')

## Keep only the rows with at least 2 non-NA values

df_student.dropna(axis=0, thresh=2)

## Define columns to erase missing values

df_student.dropna(subset=['Score', 'Grades'])

## Using inplace

df_student.dropna(inplace=True)

df_student
