==> pandas-basic-functions/32_fill_na.ipynb <==
# Filling records that have missing values with `fillna`

`DataFrame.fillna(value=None, method=None, axis=None, inplace=False, limit=None, downcast=None)`

value:
- A value that is used to fill the null values
- Can be scalar, dict, series, DataFrame

method:
- pad / ffill (front fill) : use last value as next value (top --> bottom) <br>
- bfill (back fill) : use value in front as a value backwards (bottom --> top) <br>


import pandas as pd

df = pd.read_csv("./datasets/df_na.csv")

df

## 1. Fill all NaN values with zeros

df.fillna(0)

## 2. Fill all NaN values with mean value

df['Score'].fillna(df['Score'].mean())

df_cities = pd.read_csv("./datasets/cities_temp.csv", na_values="")

df_cities

# df_cities['Temp (C)'] = df_cities['Temp (C)'].fillna(df_cities['Temp (C)'].mean())
df_cities['Temp (C)'].fillna(df_cities['Temp (C)'].mean())

# df_cities

df.fillna({'Score' : 95, 'Grades' : 'AA'})

## 3. Front fill (top bottom)

df_cities['Temp (C)'].fillna(method='ffill')

## 4. Back fill (bottom top)

df_cities['Temp (C)'].fillna(method='bfill')

## 5. Axis

df_cities

Fill values by rows

df_cities['Temp (C)'].fillna(method='ffill', axis=0)

Fill values by rows (bottom top)

df_cities['Temp (C)'].fillna(method='bfill', axis=0)

Fill values by columns (left to right)

df_cities.fillna(method='ffill', axis=1)

Fill values by columns (right to left)

df.fillna(method='bfill', axis=1)

## 6. Inplace

Make changes to the dataframe permanently

df_cities['Temp (C)'].fillna(method='ffill', inplace=True)

df_cities

## 7. Limit

limiting the number of missing values to be filled from top to bottom

df_cities_2 = pd.read_csv("./datasets/cities_temp.csv")

df_cities_2

Fill the first NaN element only

df_cities_2.fillna(method='ffill', limit=1)

## 8. Downcast

- takes a dict or string 'infer' <br>
- copies the values from previous state to the next

df_down = pd.DataFrame({'a' : [1, None]})

df_down

df_down_1 = df_down.fillna(0, downcast='infer')

df_down_1

df_down_2 = df_down.fillna({'a' : 0}, downcast='infer')

df_down_2
