==> pandas-basic-functions/33_replace.ipynb <==
# Replace

`df.replace(
    to_replace = None,
    value = None,
    inplace False
    limit = None,
    regex = False,
    method = 'pad'
)`

- to_replace : str, regex, dict, series, int, float, None<br>
  Finding values that will be replaced<br>

- value : scalar, dict, list, str, regex, default, None<br>
  Value to replace any values<br>

import pandas as pd
import numpy as np

df_students = pd.read_csv("./datasets/df_na.csv")

df_students

## Replace 'Paul' with 'Alexander'

df_students.replace(to_replace='Paul', value='Alexander')

# replace directly
df_students.replace('Paul', 'Alexander')

## Replace 'Paul' and 'Aaron' with 'Alexander'

df_students.replace(to_replace=['Paul', 'Aaron'], value='Alexander')

# Replace directly
df_students.replace(['Paul', 'Aaron'], 'Alexander')

## Replace 'NaN' with 78

df_students.replace(np.nan, 78)

## Replace NaN in 'Score' with 78 and NaN in 'Grades' with BC

# using dictionary
df_students.replace({'Score' : {np.nan : 78},
                    'Grades' : {np.nan : 'BC'}})

## Replace 'not available', np.nan, -99, in 'Grades' to 'FF'

df_students_min = pd.read_csv("./datasets/df_na_nav_min.csv")

df_students_min

df_students_min.replace({'Grades' : ['not available', np.nan, '-99']}, 'BC', inplace=True)
# add inplace=True to make changes to the dataframe permanently

## Replace Grades values: <br>
   BC : Remedial <br>
   AB : Very Good <br>
   AA : Excellent <br>
   BA : Good <br>

df_students_min.replace(['BC', 'AB', 'AA', 'BA'],
                        ['Fail', 'Very Good', 'Excellent', 'Good'])

## Replace NaN values with ffill

df_students_min

df_students_min.replace(np.nan, method='ffill')
# df_students_min.replace(np.nan, method='ffill', limit=1)

## Clean data in 'Score' using regex

df_regex = pd.read_csv("./datasets/df_na_regex.csv")

df_regex

df_regex.replace({"Score": '[A-Za-z]'}, " ", regex=True)

type(df_regex.loc[2, 'Score'])

## Replace NaN values in 'Score' with 0

df_zero = pd.read_csv("./datasets/df_na_nav.csv")

df_zero

df_zero.replace({'Score': np.nan}, 0)

## Replace NaN value (first value only) of 'Grades' with 'FF' using `at`

Access a single value for a row/column label pair

df_zero

df_zero.at[0, 'Grades'] = 'AA'

df_zero

## Replace NaN value (first value only) of 'Score' column with 0 using `iat`

iat accesses a single value for a row/column pair by integer position

df_zero.iat[1, 1] = 0

df_zero

## Replace NaN value (first value only) of 'Score' column with 0 using `at` 

Accesses a single value for a row/colum label

df_zero.at[5, 'Score'] = 0

df_zero

df_zero.at[4, 'Grades'] = np.nan

df_zero

## Replace consecutive NaN values of 'Grades' with 'AB' using `iloc`

`iloc` accesses a group of rows and columns by integer position

df_zero

df_zero.iloc[3:5, 2] = 'AB'

df_zero

## Replace consecutive NaN values of 'Grades' column with 'AB' using `loc`

Accesses a group of rows and columns by labels

df_cons = pd.read_csv("./datasets/df_na_cons.csv")

df_cons

df_cons.loc[3:4, 'Grades'] = 'AB'

df_cons
