==> pandas-basic-functions/34_interpolate.ipynb <==
# Pandas interpolate

Handling missing values using `interpolate` function

import pandas as pd
import numpy as np

`DataFrame.Interpolate()`

syntax:

`DataFrame.Interpolate(method='linear',`<br>
                        `axis=0,`<br>
                        `limit=None,`<br>
                        `inplace=False,`<br>
                        `limit_direction='forward',`<br>
                        `limit_area=None,`<br>
                        `downcast=None, **kwargs)`

parameters:

method : {'linear', 'time', 'index', 'values', 'nearest', 'zero', 'slinear', 'quadratic', 'cubic', 'barycentric', 'krogh', 'polynomial', 'spline', 'piecewise_polynomial', 'from_derivatives', 'pchip'}

axis : 
{{0 or 'index', 1 or 'columns', None}}, default None

limit : Maximum number of consecutive NaNs to fill. Must be greater than 0.

limit_direction : {‘forward’, ‘backward’, ‘both’}, default ‘forward’

limit_area : None (default) no fill restriction. 

inside :  Only fill NaNs surrounded by valid values (interpolate).

outside : Only fill NaNs outside valid values (extrapolate). 

If limit is specified, consecutive NaNs will be filled in this direction.

inplace

df_temp = pd.read_csv("./datasets/cities_temp_inter.csv")

df_temp

## Basic interpolation

df_temp.interpolate()

df_temp.interpolate(method='nearest')

## Using date

Converting a column's data type to datetime to perform time interpolation

df_temp.info()

Convert the column to datetime

df_temp['Date'] = df_temp['Date'].astype('datetime64[ns]')

Check each column's dtype

df_temp.dtypes

Check dataframe

df_temp

Set the date column as index

df_temp = df_temp.set_index(df_temp['Date'])

df_temp

Drop the date column

df_temp_drop = df_temp.drop('Date', axis=1)

df_temp_drop

Interpolate the temperature based on time

df_temp_drop.interpolate(method='time')

Using index

df_temp_drop

df_temp_drop.interpolate(method='index')

Notice the interpolation results using index are the same with time because we set the date column as an index of the dataframe

Let's actually compare with different index

df_temp = df_temp_drop.reset_index()

df_temp

df_temp['Temp (C)'].interpolate(method='piecewise_polynomial')

## Interpolate to nearest values

df_temp_drop.interpolate(method='nearest')

## Plynomial interpolation

df_temp = pd.read_csv("./datasets/cities_temp_inter.csv")

df_temp

df_temp.interpolate(method='polynomial', order=1)

df_temp.interpolate(method='spline', order=1)

## Specifying the axis

Interpolate rowwise

df_temp.interpolate(axis=0)

Interpolate columnwise

df_temp.interpolate(axis=1)

It returns an error because we don't have a previous numeric column as interpolation reference

## Set limit

Maximum number of consecutive NaNs to fill. Must be greater than zero

df_temp.interpolate(limit=2)

## Set limit direction

limit_direction : {‘forward’, ‘backward’, ‘both’}, default ‘forward’

df_temp.interpolate(limit_direction='forward')

## Limit area

limit_area : None (default) no fill restriction. 

inside :  Only fill NaNs surrounded by valid values (interpolate).

outside : Only fill NaNs outside valid values (extrapolate). 

If limit is specified, consecutive NaNs will be filled in this direction.

inplace

df_temp.interpolate(limit_area='inside')

## Inplace

Make changes and save the changes to the current dataframe

df_temp.interpolate(method='krogh', inplace=True)

df_temp

The dataframe has been changed
