==> pandas-basic-functions/35_grouping_concept.ipynb <==
# Grouping without using `groupby`

Groupby consists:

1. Split
2. Apply
3. Combine

import pandas as pd

city_temp = {'City' : ['Anaheim', 'Baltimore', 'Clearwater',
                       'Anaheim', 'Baltimore', 'Clearwater'],
             'Temp (C)' : [31, 21, 26, 30, 22, 27]}

df_city_temp = pd.DataFrame(city_temp)

df_city_temp

## 1. Split

Split the data into different groups by criteria (city)

# Anaheim
df_anaheim = df_city_temp[df_city_temp['City'] == 'Anaheim']

df_anaheim

df_baltimore = df_city_temp[df_city_temp['City'] == 'Baltimore']

df_baltimore

df_clearwater = df_city_temp[df_city_temp['City'] == 'Clearwater']

df_clearwater

## 2. Apply

Perform calculation for each group

anaheim_avg = df_anaheim['Temp (C)'].mean()

anaheim_avg

baltimore_avg = df_baltimore['Temp (C)'].mean()

baltimore_avg

clearwater_avg = df_clearwater['Temp (C)'].mean()

clearwater_avg

## 3. Combine

Combine the values based on group (city)

city_temp_combined = pd.DataFrame({
    'City' : ['Anaheim', 'Baltimore', 'Clearwater'],
    'Temp (C)' : [anaheim_avg, baltimore_avg, clearwater_avg] 
})

city_temp_combined

## 4. Complete grouping using `groupby`

# method 1
df_city_temp.groupby('City')['Temp (C)'].mean()

# method 2
df_city_temp['Temp (C)'].groupby(df_city_temp['City']).mean()
