==> pandas-basic-functions/36_grouping_groupby.ipynb <==
# Grouping using `groupby`

import pandas as pd

city_temp = {'City' : ['Anaheim', 'Baltimore', 'Clearwater',
                       'Anaheim', 'Baltimore', 'Clearwater'],
             'Temp (C)' : [31, 21, 26, 30, 22, 27],
             'Humidity':[34, 24, 23, 24, 20, 26]}

df_city_temp = pd.DataFrame(city_temp)

df_city_temp

## 1. Display Average Temperature.

df_city_temp['Temp (C)'].mean()

## 2. Display Average Temperature by City

df_city_temp.groupby('City')['Temp (C)'].mean()

df_city_temp['Temp (C)'].groupby(df_city_temp['City']).mean()

## 3. Display All The Groups (Iterate Over The Groups)

cities = df_city_temp.groupby('City')

print(cities)

print("")

for key, value in cities:
    print(key)
    print(value)
    print()

## 4. Display The Indices In Each Group.

# cities = df_city_temp.groupby('City')
cities.groups

## 5. Display The First Entry In All The Groups Formed.

# cities = df_city_temp.groupby('City')
cities.first()

# cities = df_city_temp.groupby('City')
cities.last()

## 6. Count The Number of Entries In Each Group.

# cities = df_city_temp.groupby('City')
cities.size()

## 7. Display A Specific Group Out of All The Groups.

# cities = df_city_temp.groupby('City')
cities.get_group('Baltimore')

## 8. Apply Functions To Groups [e.g. min, max and mean]

# cities = df_city_temp.groupby('City')
cities.min()

# cities = df_city_temp.groupby('City')
cities.max()

# cities = df_city_temp.groupby('City')
cities.mean()

# cities = df_city_temp.groupby('City')
mean_temp = df_city_temp.groupby('City')['Temp (C)'].mean()

mean_temp

## 9. Display Average Temperature CityWise With Prefix Mean_Temp_.

# cities = df_city_temp.groupby('City')
cities['Temp (C)'].mean().add_prefix('Avg_Temp_')

cities['Humidity'].mean().add_prefix('Avg_Humidity_')

## 10. Display Descriptive Statistics GroupWise.

# cities = df_city_temp.groupby('City')
cities.describe()

## 11. Perform Statistical Computations Using agg() Function.

# cities = df_city_temp.groupby('City')
cities.agg(['min', 'max', 'mean'])

cities['Temp (C)'].agg(['min', 'max', 'mean'])

cities['Humidity'].agg(['min', 'max', 'mean'])

## 12. Create A Function To Get The Statistics GroupWise.

# cities = df_city_temp.groupby('City')

def func(city):
    return {'Min' : city.min(),
            'Mean': city.mean()}

city_temp_apply = df_city_temp.groupby('City').apply(func)

city_temp_apply

## 13. Nesting grouping values

nested_grouping = df_city_temp.groupby(['City', 'Temp (C)'])

nested_grouping

for key, value in nested_grouping:
    print(key)
    print(value)
    print()
