==> pandas-basic-functions/37_html.ipynb <==
# Pandas HTML - Read tables from web

import pandas as pd

cities_url = "https://en.wikipedia.org/wiki/List_of_United_States_cities_by_population"

cities_data = pd.read_html(cities_url)

type(cities_data)

len(cities_data)

cities_data[2]

len(cities_data[2])
