==> pandas-basic-functions/38_merge_dataframes.ipynb <==
# Merge Pandas Dataframes

import pandas as pd

data1 = {'key' : ['k0', 'k1', 'k2', 'k3', 'k4'],
         'Name': ['Paula', 'Alexis', 'Jordan', 'Trent', 'Ulysses']}

data2 = {'key' : ['k0', 'k1', 'k2', 'k3', 'k4'],
         'Score': [96, 98, 94, 92, 92]}

df1 = pd.DataFrame(data1)

df2 = pd.DataFrame(data2)

df1

df2

## Merge left using on

Always remember that the reference column as a value of `on` must be the exact same in both dataframes

We can pass label or list as the parameter

pd.merge(left=df1, right=df2, on='key')

pd.merge(df1, df2, on='key')

pd.merge(df2, df1, on='key')

Label : single key <br>
List : multiple keys

data1 = {'key1' : ['k0', 'k1', 'k2', 'k3', 'k4'],
         'key2' : ['k0', 'k1', 'K0', 'k1', 'k2'],
         'Name': ['Paula', 'Alexis', 'Jordan', 'Trent', 'Ulysses']}

data2 = {'key1' : ['k0', 'k1', 'k2', 'k3', 'k4'],
         'key2' : ['k0', 'k1', 'K0', 'k1', 'k2'],
         'Score': [96, 98, 94, 92, 92]}

df1 = pd.DataFrame(data1)

df2 = pd.DataFrame(data2)

df1

df2

pd.merge(df1, df2, on=['key1', 'key2'])

## Inner merge/join

how ('left', 'right', 'outer', 'inner')

data1 = {'key' : ['k0', 'k1', 'k2', 'k5', 'k6'],
         'Name': ['Paula', 'Alexis', 'Jordan', 'Trent', 'Ulysses']}

data2 = {'key' : ['k0', 'k1', 'k2', 'k3', 'k4'],
         'Score': [96, 98, 94, 92, 92]}

df1 = pd.DataFrame(data1)

df2 = pd.DataFrame(data2)

df1

df2

pd.merge(df1, df2, on="key")

pd.merge(df1, df2, on='key', how='inner')

pd.merge(df1, df2, on='key', how='outer')

pd.merge(df1, df2, on='key', how='left')

pd.merge(df1, df2, on='key', how='right')

## Indicator parameter

Indocator (bool, string)

pd.merge(df1, df2, on='key', how='outer', indicator=True)

pd.merge(df1, df2, on='key', how='outer', indicator='Source')

## `left_index` and `right_index`

data1 = {'key' : ['k0', 'k1', 'k2', 'k3', 'k4'],
         'Name': ['Paula', 'Alexis', 'Jordan', 'Trent', 'Ulysses']}

data2 = {'key' : ['k5', 'k6', 'k7', 'k8', 'k9'],
         'Score': [96, 98, 94, 92, 92]}

df1 = pd.DataFrame(data1)

df2 = pd.DataFrame(data2)

df1

df2

pd.merge(df1, df2)

pd.merge(df1, df2, left_index=True, right_index=True)

## suffix

data1 = {'key' : ['k0', 'k1', 'k2', 'k3', 'k4'],
         'Name': ['Paula', 'Alexis', 'Jordan', 'Trent', 'Ulysses']}

data2 = {'key' : ['k0', 'k1', 'k2', 'k3', 'k4'],
         'Name': ['Paula', 'Alexis', 'Jordan', 'Trent', 'Ulysses']}

df1 = pd.DataFrame(data1)

df2 = pd.DataFrame(data2)

df1

df2

pd.merge(df1, df2, on='key')

pd.merge(df1, df2, on='key', suffixes=('_leftDF', '_rightDF'))
