==> pandas-basic-functions/39_concat_series.ipynb <==
# Dataframe concatenations

import pandas as pd

series1 = pd.Series([1, 2, 3, 4, 5])
series2 = pd.Series(['A', 'B', 'C', 'D', 'E'])

series1

series2

## Concatenate vertically

pd.concat([series1, series2])

pd.concat([series1, series2], axis=0)

## Concatenate horizontally

pd.concat([series2, series1], axis=1)

## Concatenate series with different composition

series3 = pd.Series([1, 2, 3, 4, 5])
series4 = pd.Series(['A', 'B', 'C', 'D'])

series3

series4

pd.concat([series3, series4])

Concatenate different values horizontally

pd.concat([series3, series4], axis=1)

## Concatenate series vartically using `concat()` and reset the index

series5 = pd.Series([1, 2, 3, 4, 5])
series6 = pd.Series(['A', 'B', 'C', 'D', 'E'])

series5

series6

pd.concat([series5, series6], axis=0, ignore_index=True)

## Concatenate dataframes horizontally using `concat()`

data1 = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82]
}

data2 = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Aurora', 'Madison', 'Paxton'],
    'Score': [98, 89, 99, 87, 82, 83, 90]
}

df1 = pd.DataFrame(data1, index=[0, 1, 2, 3, 4, 5, 6])
df2 = pd.DataFrame(data2, index=[0, 1, 2, 14, 15, 16, 17])

df1

df2

pd.concat([df1, df2], axis=1)

pd.concat([df1, df2], axis=0)

pd.concat([df1, df2], axis=0, ignore_index=True)

## Join

pd.concat([df1, df2], axis=1, join='inner')
