==> pandas-basic-functions/40_concat_dataframes.ipynb <==
# Dataframes concatenations - Part 2

import pandas as pd

data1 = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82]
}

data2 = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Aurora', 'Madison', 'Paxton'],
    'Score': [98, 89, 99, 87, 82, 83, 90]
}

df1 = pd.DataFrame(data1, index=[0, 1, 2, 3, 4, 5, 6])
df2 = pd.DataFrame(data2, index=[0, 1, 2, 14, 15, 16, 17])

## Concat vertically

pd.concat([df1, df2], axis=0)

## Concat horizontally

pd.concat([df1, df2], axis=1)

## Ignore index

pd.concat([df1, df2], axis=0, ignore_index=True)

## Concatenate two different columns

data3 = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82]
}

data4 = {
    'Grades': ['AA', 'AB', 'AA', 'AB', 'AB', 'AC', 'AC']
}

df3 = pd.DataFrame(data3)
df4 = pd.DataFrame(data4)

df3

df4

pd.concat([df3, df4], axis=1)

## Concatenate multiple dataframes

data5 = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82]
}

data6 = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Aurora', 'Madison', 'Paxton'],
    'Score': [98, 89, 99, 87, 82, 83, 90]
}

data7 = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Aurora', 'Madison', 'Paxton'],
    'Score': [98, 89, 99, 87, 82, 83, 90]
}

df5 = pd.DataFrame(data5)
df6 = pd.DataFrame(data6)
df7 = pd.DataFrame(data7)

Vertically

pd.concat([df5, df6, df7], axis=0)

Horizontally

pd.concat([df5, df6, df7], axis=1)

## Concatenate dataframes with series

data8 = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Aurora', 'Madison', 'Paxton'],
    'Score': [98, 89, 99, 87, 82, 83, 90]
}

df8 = pd.DataFrame(data8)

s1 = pd.Series(['AA', 'AB', 'AA', 'AB', 'AB', 'AC', 'AC'], name='Grades')

df8

s1

Vertically

pd.concat([df8, s1])

Horizontally

pd.concat([df8, s1], axis=1)
