==> pandas-basic-functions/41_concat_parameters.ipynb <==
# Dataframes concatenations - Part 3 - Parameters

Concat parameters:

axis=0, <br>
ignore_index (bool) = False, <br>
join='outer', <br>
keys=None, <br>
names=None, <br>
verify_integrity (bool) = False, <br>
sort (bool) = False, <br>

import pandas as pd

data1 = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton', 'Madison', 'Aurora'],
    'Score': [98, 89, 99, 87, 90, 83, 82]
}

data2 = {
    'Name' : ['Robert', 'Craig', 'Frank', 'Jordyn', 'Isabel'],
    'Score': [98, 89, 99, 97, 97]
}

df1 = pd.DataFrame(data1)
df2 = pd.DataFrame(data2)

df1

df2

## Basic concat

pd.concat([df1, df2], axis=0)

## Using keys parameter

Vertically

pd.concat([df1, df2], axis=0, keys=['DF1', 'DF2'])

Horizontally

pd.concat([df1, df2], axis=1, keys=['DF1', 'DF2'])

## Store concatenated dataframe to a variable and make hierarchical dataframe

data = pd.concat([df1, df2], axis=0, keys=['DF1', 'DF2'])

data

## Accessing the values in the dataframe by rows/columns

data.loc['DF1']

Or

data.loc['DF1', :]

Or

data.loc['DF2']

data.loc['DF2', :]

## Add names to hierarchical index

pd.concat([df1, df2], keys=['DF1', 'DF2'], names=['Level_1', 'Level_2'])

## Verify integrity parameter (avoid duplicate indices)

`verify_integrity=False` (bool, default=False) 

data1 = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton'],
    'Score': [98, 89, 99, 87, 90]
}

data2 = {
    'Name' : ['Robert', 'Craig', 'Frank', 'Jordyn', 'Isabel'],
    'Score': [98, 89, 99, 97, 97]
}

df1 = pd.DataFrame(data1, index=[i for i in range(5)])
df2 = pd.DataFrame(data2, index=[i for i in range(5)])

df1

df2

Verify integrity (checks if a dataframe contans duplicated indices) <br>

For the dataframe above we have duplicated indices such that in df1 we have 0, 1, 2, 3, 4 <br>
and df2 we also have 0, 1, 2, 3, 4, as indices. <br>

Because of this, `verify_integrity=True` should give an error

pd.concat([df1, df2], axis=0, verify_integrity=True)

if you don't want the error, then you should set different values as indices to different datasets

Similar thing will happen if we set the axis to 1 (`axis=1`). <br>
Because in both dataframes we have the exact same column names as horizontal indices. ('Names' and 'Score')

pd.concat([df1, df2], axis=1, verify_integrity=True)

See, we get an error

## Sort parameter

Sort columns if we are adding rows, and sort rows if we are adding columns

data1 = {
    'Name' : ['Paul', 'Aaron', 'Krista', 'Veronica', 'Paxton'],
    'Score': [98, 89, 99, 87, 90]
}

data2 = {
    'Name' : ['Robert', 'Jordyn', 'Isabel'],
    'Grades': ['AA', 'AA', 'AA']
}

df1 = pd.DataFrame(data1, index=[i for i in range(5)])
df2 = pd.DataFrame(data2, index=[i for i in range(3)])

df1

df2

pd.concat([df1, df2], axis=0)

Again, we are sorting columns if we are adding rows, and vice versa

Using `axis=0`, we get:

pd.concat([df1, df2], axis=0, sort=True)

Using `axis=1`, we get:

pd.concat([df1, df2], axis=1, sort=True)

Our dataframe is sorted by index values
