==> pandas-basic-functions/42_sampling_dataset.ipynb <==
# Fetching and sampling random values from a dataset

`df.sample(` <br>
    `n=None,` <br>
    `frac=None,` <br>
    `replace=False,` <br>
    `weights=None,` <br>
    `random_state=None,` <br>
    `axis=None,` <br>
`)`

Note : `n` and `frac` cannot be used together

import pandas as pd

df_iris = pd.read_csv("./datasets/iris.csv")

df_iris.head()

df_iris.info()

df_iris['Species'].value_counts()

# Syntax and parameters

`df.sample(` <br>
    `n=None,` <br>
    `frac=None,` <br>
    `replace=False,` <br>
    `weights=None,` <br>
    `random_state=None,` <br>
    `axis=None,` <br>
`)`

Note : `n` and `frac` cannot be used together

## 1. Fetch 5 random samples from the dataset

df_iris.sample(
    n=5
)

## 2. Fetch 10% of data from the dataset (frac parameter)

df_iris.sample(frac=0.10) 

## 3. Random state parameter

Give the same output samples every run

df_iris.sample(n=3, random_state=200)

## 4. Replace parameeter

Repeat samples occurence in for every next run

data_sample = {
    'Name' : ['Robert', 'Craig', 'Frank', 'Jordyn', 'Isabel'],
    'Score': [98, 89, 99, 97, 97]
}

df_sample = pd.DataFrame(data_sample)

df_sample

df_sample.sample(n=2, replace=False)

## 5. Axis parameter

df_iris.head()

Getting 5 random rows (`axis=0`) from the dataset 

df_iris.sample(n=5, axis=0)

Getting 2 random columns (`axis=1`) from the dataset

df_iris.sample(n=2, axis=1)

## 6. Weights parameter

data_sample = {
    'Name' : ['Robert', 'Craig', 'Frank', 'Jordyn', 'Isabel'],
    'Score': [98, 89, 99, 97, 97]
}

df_sample = pd.DataFrame(data_sample)

df_sample

Set probability values (weighting) for sampling dataset

# Set weights (probability)
w = [0.25, 0.45, 0.8, 0.34, 0.2]

df_sample.sample(n=3, weights=w)

Note: Weights and axis (rows) to be sampled must be of same length

## 7. Pandas series

df_sample['Name']

type(df_sample['Name'])

df_sample['Name'].sample(n=2)
